feat: Add `.from_tree` constructor
The `Node.from_tree` constructor directly creates a tree from any `Node`
subclass given a known `(value, children)` tuple. This is useful for
building a tree from a flat representation.

Subclasses of `Node` can use `__init__` to build the tree without
existing `(value, children)`. This is common where a tree structure is
predefined for a class (for example a `Quadrilateral` primitive is known
to have 4 lines as children).