bug: Modify `transform_sum` to better reflect sum
Previous implementation had identical primitives and parameters to each
primitive and parameter for the summed constructions. It make more sense
that the sum construction should accept prims and params for summed
construction then return the sum. That is, `prims` for the sum
construction should contain prim arguments for both summed
constructions.