bug: Update `transform_scalar...` and `transform_partial` class inheritance
The old implementations would return transformed construction classes that
didn't inherit from the input constructions. This meant that an input
`Scalar` would no longer be a scalar when raised to a power, for
example. This prevented `transform_scalar_div` from working since it
multiplies by the inverse of a scalar to do the operation.